net: lib: nrf_cloud: move header file into /net
Move nrf_cloud.h header file from /include to /include/net.
