+++ samples/nrf9160/lte_ble_gateway/src/main.c
@@ -10,7 +10,7 @@
 #include <drivers/gps.h>
 #include <drivers/sensor.h>
 #include <console/console.h>
-#include <nrf_cloud.h>
+#include <net/nrf_cloud.h>
 #include <dk_buttons_and_leds.h>
 #include <modem/lte_lc.h>
 #include <power/reboot.h>
